requirements.txt
streamlit==1.30.0
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
ffmpeg-python==0.2.0

